moved exceptions and mappers to separate packages to clean it up
