+++ src/main/java/at/aau/ase/cl/api/interceptor/exceptions/EmailAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package at.aau.ase.cl.api.interceptor;
+package at.aau.ase.cl.api.interceptor.exceptions;
 
 public class EmailAlreadyExistsException extends RuntimeException {
     public EmailAlreadyExistsException(String message) {
