+++ src/main/java/at/aau/ase/cl/api/interceptor/exceptions/InvalidPasswordException.java
@@ -1,4 +1,4 @@
-package at.aau.ase.cl.api.interceptor;
+package at.aau.ase.cl.api.interceptor.exceptions;
 
 public class InvalidPasswordException extends RuntimeException {
     public InvalidPasswordException(String message) {
