+++ src/main/java/at/aau/ase/cl/api/interceptor/exceptions/NotFoundException.java
@@ -1,4 +1,4 @@
-package at.aau.ase.cl.api.interceptor;
+package at.aau.ase.cl.api.interceptor.exceptions;
 
 public class NotFoundException extends RuntimeException {
     public NotFoundException(String message) {
