+++ src/main/java/at/aau/ase/cl/api/interceptor/exceptions/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package at.aau.ase.cl.api.interceptor;
+package at.aau.ase.cl.api.interceptor.exceptions;
 
 public class UserNotFoundException extends RuntimeException {
     public UserNotFoundException(String message) {
