+++ src/main/java/at/aau/ase/cl/api/interceptor/mapper/IllegalArgumentExceptionMapper.java
@@ -1,5 +1,6 @@
-package at.aau.ase.cl.api.interceptor;
+package at.aau.ase.cl.api.interceptor.mapper;
 
+import at.aau.ase.cl.api.interceptor.ErrorResponse;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.ext.ExceptionMapper;
 import jakarta.ws.rs.ext.Provider;
