+++ src/main/java/at/aau/ase/cl/api/interceptor/mapper/NotFoundExceptionMapper.java
@@ -1,5 +1,7 @@
-package at.aau.ase.cl.api.interceptor;
+package at.aau.ase.cl.api.interceptor.mapper;
 
+import at.aau.ase.cl.api.interceptor.ErrorResponse;
+import at.aau.ase.cl.api.interceptor.exceptions.NotFoundException;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.ext.ExceptionMapper;
