+++ src/main/java/at/aau/ase/cl/service/UserService.java
@@ -1,9 +1,9 @@
 package at.aau.ase.cl.service;
 
-import at.aau.ase.cl.api.interceptor.EmailAlreadyExistsException;
-import at.aau.ase.cl.api.interceptor.NotFoundException;
-import at.aau.ase.cl.api.interceptor.UserNotFoundException;
-import at.aau.ase.cl.api.interceptor.UsernameAlreadyExistsException;
+import at.aau.ase.cl.api.interceptor.exceptions.EmailAlreadyExistsException;
+import at.aau.ase.cl.api.interceptor.exceptions.NotFoundException;
+import at.aau.ase.cl.api.interceptor.exceptions.UserNotFoundException;
+import at.aau.ase.cl.api.interceptor.exceptions.UsernameAlreadyExistsException;
 import at.aau.ase.cl.model.AddressEntity;
 import at.aau.ase.cl.model.UserEntity;
 import io.quarkus.elytron.security.common.BcryptUtil;
